Trait pallet_mandate::pallet::Config

pub trait Config: Config {
    type Event: From<Event<Self>> + IsType<<Self as Config>::Event>;
    type Call: Parameter + UnfilteredDispatchable<Origin = Self::Origin> + GetDispatchInfo;
    type ExternalOrigin: EnsureOrigin<Self::Origin>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type Event: From<Event<Self>> + IsType<<Self as Config>::Event>

type Call: Parameter + UnfilteredDispatchable<Origin = Self::Origin> + GetDispatchInfo

type ExternalOrigin: EnsureOrigin<Self::Origin>

Origin that can call this module and execute sudo actions. Typically the collective module.